third_party/blink/web_tests/platform/generic/fast/forms/image/input-width-height-attributes-without-renderer-not-loaded-image-expected.txt

PASS e.width is 0
PASS e.height is 0
PASS successfullyParsed is true

TEST COMPLETE
